pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-security-gate"
version = "0.1.0"
description = "Product scaffold for a policy-enforced agent security gate with CI benchmark support."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.6",
  "PyYAML>=6.0",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.27.0",
  "httpx>=0.27.0",
  "psycopg[binary]>=3.1.0",
  "PyJWT>=2.8.0",
  "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6.0",
]

[tool.setuptools]
packages = ["adapters", "app", "approvals", "audit", "benchmark", "gateway", "scripts"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
  "integration: requires docker compose (gateway + OPA)",
]
